Initial implementation.
This version only supports HTTP basic authentication with and without caching.
In the future it would be nice to support OAuth2 authentication as well.  This
version also only supports API key definition within the Dropwizard
configuration file.  In the future it may be possible to provide a provider
for API keys so that they may be retrieved from a data store.