Refactor to use String instead of ApiKey class.
This change refactors the object that is returned after a successful
authorization operation.  Instead of an ApiKey object like was
previously returned, a java.lang.String is now just returned.  This
string is the name of the application that made the request.  This
change feels right in a few ways.  First it makes the implementation
of service methods simpler because they have less to worry about
because they're only receiving a string instead of some random
object from a library.  Secondly by returning a string instead of
the full ApiKey object we're preventing code that doesn't need to
have access to the API key secret from having direct access to it.
This strictly isn't a huge deal, but it feels more correct than
passing it around everywhere.